requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
SpeechRecognition==3.10.0
PyAudio==0.2.13
gTTS==2.3.2
requests==2.31.0
duckduckgo-search==3.9.3
PyNaCl==1.5.0
langdetect==1.0.9
pyttsx3==2.90
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0 